rhbootstrap: Add symlink sh -> rhbootstrap.sh
Most of the times rhbootstrap is git cloned on target host and thus
rhbootstrap.sh stored in rhbootstrap/ directory unser regular user and
called with sudo(8) or under root user with absolute path with
"rhbootstrap" appearing in it multiple times.

To avoid multiple "rhbootstrap" name in absolute path introduce symlink
sh that points to rhbootstrap.sh file that can be used instead of
rhbootstrap.sh name in absolute path:

  $ sudo -i ~/git/rhbootstrap/sh --help

instead of

  $ sudo -i ~/git/rhbootstrap/rhbootstrap.sh --help

previously.

Signed-off-by: Serhey Popovych <serhe.popovych@gmail.com>