rhbootstrap: Drop Nux Dextop repository support
We reference limited set of packages that aren't quite common in regular
CentOS reployment and there is no support (and guess there is no plans)
for CentOS 8.

Since there is little sense to keep support for this repo for CentOS 7
only remove support for it. Existing users if any, should now switch to
RPM Fusion that supported for both CentOS 7 and 8.

Signed-off-by: Serhey Popovych <serhe.popovych@gmail.com>