rhbootstrap: Add support for CentOS releases >= 4.0 and i386 architecture
At the time of this commit only CentOS 7 and 8 are maintained releases
with CentOS 8 supported to the end of 2021. Other releases are now
unsupported, do not get any (including security) updates and generally
shouldn't be used for new deployments.

On the other hand there might be a lot of unported applications to newer
versions of CentOS in enterprise environments still in use. Porting
these applications to new frameworks/environments either complex or
even impossible without full reimplementation.

To support such applications in modern (microserviced) environments we
need to be able to build containers for really old CentOS releases and
support (now) legacy i386 architecure. Repositories for such distros
available either as local mirror in enterprise network or, in case of
CentOS, at vault.centos.org.

Packages can be installed to chroot with help of yum(8) or dnf(8) on
CentOS 7 or 8 host; rpmdb(1) can be fixed by rpmdb_dump on host with
it's output piped to rpmdb_load running in chroot. The only exception
at the moment is 6.10 release where rpm(1) fails to work with
correct(ed) Packages database even after dump/load. That release
installed as an update from 6.9 (i.e. yum -y update in chroot).-

Only minimal installation supported at the moment with only EPEL as
external repository (available at archives.fedoraproject.org). Due to
nature of EPEL there is only guarantee that packages from it will be
installed (worked) on last release in the series (e.g. 5.11 in case of
5, 6.10 in case on 6, etc). It is also possible to do a minimal install
of specific release (e.g. 7.7 or 8.1) for currently maintained CentOS.

There is also support for CentOS 7 on i386 as part of AltArch initiative
(there is to EPEL available for it).

All yum(1) *.repo files fixed to point to correct archive repository
server (e.g. vault.centos.org and archives.fedoraproject.org) so that
yum(1) will be able to install required packages.

From security perspective @core package set installed without GPG/PGP
signature validation for installed packages (see --nogpgcheck option)
that might (should) be concern in security sensitive deployments. It
is neither new nor introduced with this change and SHOULD be addressed
in future releases of rhbootstrap. For now it is assumed that packages
fetched from secure enterprise mirror of public repository.

Tested on CentOS 7 and 8 as build host for CentOS >= 4.0 containers.

Signed-off-by: Serhey Popovych <serhe.popovych@gmail.com>