rhbootstrap: Do not override repo.mirrorlist= with file:///dev/null
It is tested with CentOS 7/8 that repo.baseurl= takes precedence over
repo.mirrorlist= at least when given on command line with --setopt=....
Specifying --setopt=repo.mirrorlist= does not work on CentOS 7.

On the other hand, when specified, may cause failures on Fedora.
Remove for now to fix installation on Fedora.

Fixes: 1a617f2593d4 ("rhbootstrap: Automatically pick repos on host to use for initial setup")
Signed-off-by: Serhey Popovych <serhe.popovych@gmail.com>