rhbootstrap: Implement safe_curl() helper to restrict download size
It is unsafe to fetch from network unrestricted amount of data as this
may have DoS impact on build system. Unfortunately curl(1) does not
provide reliable builtin mechanisms to force limits. To implement that
pipe data from curl(1) standard output to head(1) standard input and
force it to quit after reading no more than K bytes given with -c.

This is reliable as far as curl(1) terminates on SIGPIPE.

Signed-off-by: Serhey Popovych <serhe.popovych@gmail.com>