rhbootstrap: Rename systemd_edit() to systemctl_edit() and put it global
While moving helper to global contect it was realized that helper name
does not reflect command name it is build upon. Rename helper to
systemctl_edit() to address this note.

Put helper into global contect to make it available to entire script
since there will be more places where it can be used.

Signed-off-by: Serhey Popovych <serhe.popovych@gmail.com>