rhbootstrap: Avoid installing packages weak dependencies
It was noted that installation of virt-manager(1) causes installation of
libvirt and qemu-kvm packages listed for that package as weak deps. This
is unexpected on pure desktop systems that not intended to be used as
virtualization hosts.

Add --setopt=install_weak_deps=${install_weak_deps} option to all yum(1)
calls to prevent weak dependencies installation.

While there it was observed that other --setopt= respected only for
base system installs, not for rest of packages installed explicitly or
via profile configuration. To address this instroduce and use set of
helpers that add required options to yum(1).

Signed-off-by: Serhey Popovych <serhe.popovych@gmail.com>