rhbootstrap: Declare Fedora support
It was supported onofficially for a while, however only Fedora versions
from 10 to 32 until now due to rpm database format change to SQLite.

Now rpm database format migration implemented and more options added to
control packages in various releases of Fedora we can declare support of
Fedora in README file with this commit.

Signed-off-by: Serhey Popovych <serhe.popovych@gmail.com>