rhbootstrap: Update CentOS 7 repo kickstart option to point to the vault
This release is EoL since June 30, 2024 thus mirrorlist.centos.org isn't
available anymore as well as repositories removed from most of mirror
sites worldwide.

It is however might be necessary to intall that release for some reasons
and it is still possible by using valut.centos.org instead of mirrorlist
in URLs.

Signed-off-by: Serhey Popovych <serhe.popovych@gmail.com>